More refactoring and cleaning
