+++ examples/mnist-keras/client/entrypoint
@@ -1,8 +1,9 @@
 #!./.mnist-keras/bin/python
 import json
 import os
+import random
 
-import docker
+#import docker
 import fire
 import numpy as np
 import tensorflow as tf
@@ -14,10 +15,10 @@ NUM_CLASSES = 10
 
 def _get_data_path():
     # Figure out FEDn client number from container name
-    client = docker.from_env()
-    container = client.containers.get(os.environ['HOSTNAME'])
-    number = container.name[-1]
-
+    #client = docker.from_env()
+    #container = client.containers.get(os.environ['HOSTNAME'])
+    #number = container.name[-1]
+    number = random.randint(0, 1)
     # Return data path
     return f"/var/data/clients/{number}/mnist.npz"
