+++ fedn/fedn/network/combiner/server.py
@@ -16,7 +16,7 @@ from fedn.common.net.grpc.server import Server
 from fedn.common.storage.s3.s3repo import S3ModelRepository
 from fedn.common.tracer.mongotracer import MongoTracer
 from fedn.network.combiner.modelservice import ModelService
-from fedn.network.combiner.roundcontrol import RoundControl
+from fedn.network.combiner.round import CombinerRound
 
 VALID_NAME_REGEX = '^[a-zA-Z0-9_-]*$'
 
@@ -111,7 +111,7 @@ class Combiner(rpc.CombinerServicer, rpc.ReducerServicer, rpc.ConnectorServicer,
         self.tracer = MongoTracer(
             config['statestore']['mongo_config'], config['statestore']['network_id'])
 
-        self.control = RoundControl(
+        self.control = CombinerRound(
             self.id, self.repository, self, self.modelservice)
         threading.Thread(target=self.control.run, daemon=True).start()
